requirements.txt
click==8.1.7
Flask==3.1.0
mysql-connector-python==9.1.0
Requests==2.32.3
setuptools==75.5.0
waitress==3.0.2
